Optimizations and makefile tweaks.
A bit of code optimizations and makefile tweaks.

Makefile tweaking was done on ARMv7 target. Different architectures may
require further fine-tuning. Base speed up on ARMv7 was from 7m42s to
7m32s execution time with elimination of -mword-realocations, various
gcc switches that are listed in makefile yelded final result of as low
as 7m26s execution time.

+++ src/zopfli/crc32.c
@@ -30,7 +30,7 @@ static int crc_table_computed = 0;
 /* Makes the table for a fast CRC. */
 static void MakeCRCTable(void) {
   unsigned long c;
-  int n, k;
+  size_t n, k;
   for (n = 0; n < 256; n++) {
     c = (unsigned long) n;
     for (k = 0; k < 8; k++) {
